Plan 9 from Bell Labs 1992-09-02

+++ port/portdat.h
@@ -732,7 +732,7 @@ extern	int	cpuserver;
 extern	char*	devchar;
 extern	Dev	devtab[];
 extern  char	eve[];
-extern	ulong	initcode[];
+extern	uchar	initcode[];
 extern	FPsave	initfp;
 extern  KIOQ	kbdq;
 extern  IOQ	lineq;
